+++ appGUI/preferences/tools/ToolsMarkersPrefGroupUI.py
@@ -12,12 +12,12 @@ if '_' not in builtins.__dict__:
     _ = gettext.gettext
 
 
-class ToolsCornersPrefGroupUI(OptionsGroupUI):
+class ToolsMarkersPrefGroupUI(OptionsGroupUI):
     def __init__(self, defaults, decimals=4, parent=None):
         # OptionsGroupUI.__init__(self, "Calculators Plugin", parent=parent)
-        super(ToolsCornersPrefGroupUI, self).__init__(self, parent=parent)
+        super(ToolsMarkersPrefGroupUI, self).__init__(self, parent=parent)
 
-        self.setTitle(str(_("Corner Markers Options")))
+        self.setTitle(str(_("Markers Options")))
         self.decimals = decimals
         self.defaults = defaults
 
@@ -64,18 +64,31 @@ class ToolsCornersPrefGroupUI(OptionsGroupUI):
         param_grid.addWidget(self.thick_label, 4, 0)
         param_grid.addWidget(self.thick_entry, 4, 1)
 
-        # Margin #
-        self.margin_label = FCLabel('%s:' % _("Margin"))
-        self.margin_label.setToolTip(
+        # Margin X #
+        self.marginx_label = FCLabel('%s X:' % _("Margin"))
+        self.marginx_label.setToolTip(
             _("Bounding box margin.")
         )
-        self.margin_entry = FCDoubleSpinner()
-        self.margin_entry.set_range(-10000.0000, 10000.0000)
-        self.margin_entry.set_precision(self.decimals)
-        self.margin_entry.setSingleStep(0.1)
+        self.marginx_entry = FCDoubleSpinner()
+        self.marginx_entry.set_range(-10000.0000, 10000.0000)
+        self.marginx_entry.set_precision(self.decimals)
+        self.marginx_entry.setSingleStep(0.1)
 
-        param_grid.addWidget(self.margin_label, 6, 0)
-        param_grid.addWidget(self.margin_entry, 6, 1)
+        param_grid.addWidget(self.marginx_label, 6, 0)
+        param_grid.addWidget(self.marginx_entry, 6, 1)
+
+        # Margin Y #
+        self.marginy_label = FCLabel('%s Y:' % _("Margin"))
+        self.marginy_label.setToolTip(
+            _("Bounding box margin.")
+        )
+        self.marginy_entry = FCDoubleSpinner()
+        self.marginy_entry.set_range(-10000.0000, 10000.0000)
+        self.marginy_entry.set_precision(self.decimals)
+        self.marginy_entry.setSingleStep(0.1)
+
+        param_grid.addWidget(self.marginy_label, 8, 0)
+        param_grid.addWidget(self.marginy_entry, 8, 1)
 
         # Length #
         self.l_label = FCLabel('%s:' % _("Length"))
@@ -87,8 +100,8 @@ class ToolsCornersPrefGroupUI(OptionsGroupUI):
         self.l_entry.set_precision(self.decimals)
         self.l_entry.setSingleStep(10 ** -self.decimals)
 
-        param_grid.addWidget(self.l_label, 8, 0)
-        param_grid.addWidget(self.l_entry, 8, 1)
+        param_grid.addWidget(self.l_label, 10, 0)
+        param_grid.addWidget(self.l_entry, 10, 1)
 
         # Drill Tool Diameter
         self.drill_dia_label = FCLabel('%s:' % _("Drill Dia"))
@@ -100,7 +113,7 @@ class ToolsCornersPrefGroupUI(OptionsGroupUI):
         self.drill_dia_entry.set_precision(self.decimals)
         self.drill_dia_entry.setWrapping(True)
 
-        param_grid.addWidget(self.drill_dia_label, 10, 0)
-        param_grid.addWidget(self.drill_dia_entry, 10, 1)
+        param_grid.addWidget(self.drill_dia_label, 12, 0)
+        param_grid.addWidget(self.drill_dia_entry, 12, 1)
 
         self.layout.addStretch()